Move class to its own Python module
